+++ installer/windows/launcher/__init__.py
@@ -1,0 +1,2 @@
+"""Windows launcher prototype package for LoreForge Lite."""
+